moved isValidChart into plugins/pkg so it can be reused by multiple plugins
