+++ cmd/kubeapps-apis/plugins/helm/packages/v1alpha1/server.go
@@ -341,7 +341,7 @@ func pageOffsetFromPageToken(pageToken string) (int, error) {
 func AvailablePackageSummaryFromChart(chart *models.Chart) (*corev1.AvailablePackageSummary, error) {
 	pkg := &corev1.AvailablePackageSummary{}
 
-	isValid, err := isValidChart(chart)
+	isValid, err := packageutils.IsValidChart(chart)
 	if !isValid || err != nil {
 		return nil, status.Errorf(codes.Internal, "invalid chart: %s", err.Error())
 	}
@@ -494,7 +494,7 @@ func (s *Server) GetAvailablePackageVersions(ctx context.Context, request *corev
 func AvailablePackageDetailFromChart(chart *models.Chart, chartFiles *models.ChartFiles) (*corev1.AvailablePackageDetail, error) {
 	pkg := &corev1.AvailablePackageDetail{}
 
-	isValid, err := isValidChart(chart)
+	isValid, err := packageutils.IsValidChart(chart)
 	if !isValid || err != nil {
 		return nil, status.Errorf(codes.Internal, "invalid chart: %s", err.Error())
 	}
@@ -564,37 +564,6 @@ func (s *Server) hasAccessToNamespace(ctx context.Context, cluster, namespace st
 		return status.Errorf(codes.Unauthenticated, "The current user has no access to the namespace %q", namespace)
 	}
 	return nil
-}
-
-// isValidChart returns true if the chart model passed defines a value
-// for each required field described at the Helm website:
-// https://helm.sh/docs/topics/charts/#the-chartyaml-file
-// together with required fields for our model.
-func isValidChart(chart *models.Chart) (bool, error) {
-	if chart.Name == "" {
-		return false, status.Errorf(codes.Internal, "required field .Name not found on helm chart: %v", chart)
-	}
-	if chart.ID == "" {
-		return false, status.Errorf(codes.Internal, "required field .ID not found on helm chart: %v", chart)
-	}
-	if chart.Repo == nil {
-		return false, status.Errorf(codes.Internal, "required field .Repo not found on helm chart: %v", chart)
-	}
-	if chart.ChartVersions == nil || len(chart.ChartVersions) == 0 {
-		return false, status.Errorf(codes.Internal, "required field .chart.ChartVersions[0] not found on helm chart: %v", chart)
-	} else {
-		for _, chartVersion := range chart.ChartVersions {
-			if chartVersion.Version == "" {
-				return false, status.Errorf(codes.Internal, "required field .ChartVersions[i].Version not found on helm chart: %v", chart)
-			}
-		}
-	}
-	for _, maintainer := range chart.Maintainers {
-		if maintainer.Name == "" {
-			return false, status.Errorf(codes.Internal, "required field .Maintainers[i].Name not found on helm chart: %v", chart)
-		}
-	}
-	return true, nil
 }
 
 // GetInstalledPackageSummaries returns the installed packages managed by the 'helm' plugin
